pyproject.toml
[tool.poetry]
name = "calculadoramcp"
version = "0.1.0"
description = "A simple MCP"
readme = "README.md"
authors = ["Jonathan Rodríguez Alejos <jrodriguez.5716@gmail.com>"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10.18"
mcp="1.4.1"
gunicorn = "^23.0.0"
uvicorn = "^0.35.0"
azure-ai-inference = "^1.0.0b9"


[tool.poetry.group.dev.dependencies]
debugpy="1.8.8"